python-service/requirements.txt
rembg==2.0.67
flask==3.0.0
flask-cors==4.0.0
pillow>=10.0.0
